# openapi.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "Read-only API for online cinema (movies).",
        "version": "0.1.0"
    },
    "paths": {
...
        "/api/v1/writers/": {
            "get": {
                "tags": [
                    "writers"
                ],
                "summary": "Writer Collection",
                "description": "Get writers collection with related films ordered by rating.",
                "operationId": "writer_collection_api_v1_writers__get",
                "parameters": [
                    {
                        "required": false,
                        "schema": {
                            "title": "Page",
                            "minimum": 1.0,
                            "type": "integer",
                            "default": 1
                        },
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "required": false,
                        "schema": {
                            "title": "Size",
                            "minimum": 1.0,
                            "type": "integer",
                            "default": 10
                        },
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Page_PersonDetailSchema_"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/writers/{writer_id}": {
            "get": {
                "tags": [
                    "writers"
                ],
                "summary": "Actor Detail",
                "description": "Get one writer with related films ordered by rating.",
                "operationId": "actor_detail_api_v1_writers__writer_id__get",
                "parameters": [
                    {
                        "required": true,
                        "schema": {
                            "title": "Writer Id",
                            "type": "string"
                        },
                        "name": "writer_id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PersonDetailSchema"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "FilmIDMixin": {
                "title": "FilmIDMixin",
                "required": [
                    "id"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "title": "Id",
                        "type": "string"
                    }
                }
            },
...
            "HTTPValidationError": {
                "title": "HTTPValidationError",
                "type": "object",
                "properties": {
                    "detail": {
                        "title": "Detail",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        }
                    }
                }
            },
...
            "Page_PersonDetailSchema_": {
                "title": "Page[PersonDetailSchema]",
                "required": [
                    "items",
                    "total",
                    "page",
                    "size"
                ],
                "type": "object",
                "properties": {
                    "items": {
                        "title": "Items",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/PersonDetailSchema"
                        }
                    },
                    "total": {
                        "title": "Total",
                        "minimum": 0.0,
                        "type": "integer"
                    },
                    "page": {
                        "title": "Page",
                        "minimum": 1.0,
                        "type": "integer"
                    },
                    "size": {
                        "title": "Size",
                        "minimum": 1.0,
                        "type": "integer"
                    }
                }
            },
...
            "PersonDetailSchema": {
                "title": "PersonDetailSchema",
                "required": [
                    "id",
                    "full_name",
                    "film"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "title": "Id",
                        "type": "string"
                    },
                    "full_name": {
                        "title": "Full Name",
                        "type": "string"
                    },
                    "film": {
                        "title": "Film",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/FilmIDMixin"
                        }
                    }
                }
            },
...
            "ValidationError": {
                "title": "ValidationError",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "type": "object",
                "properties": {
                    "loc": {
                        "title": "Location",
                        "type": "array",
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        }
                    },
                    "msg": {
                        "title": "Message",
                        "type": "string"
                    },
                    "type": {
                        "title": "Error Type",
                        "type": "string"
                    }
                }
            }
        }
    }
}